Only bind supported interface versions
The version that the server sends with a wl_registry::global event is
the highest version that the server supports. If the client does not
specifically support that version, it must not request it, as it may
change interface semantics, or result in events being sent the client
cannot understand.